Module rendy_memory::allocator::linear

Structs

Line

LinearAllocator
Linear allocator that return memory from chunk sequentially. It keeps only number of bytes allocated from each chunk. Once chunk is exhausted it is placed into list. When all blocks allocated from head of that list are freed, head is freed as well.

LinearBlock
Memory block allocated from LinearAllocator

LinearConfig
Config for LinearAllocator.