Module rendy_memory::mapping::range

Functions

mapped_fitting_range
Get sub-range of memory mapping. range and fitting is in memory object space. ptr points to the range.start offset from memory origin. returns pointer to fitting.start offset from memory origin if fitting is contained in range.

mapped_slice⚠
Safety

mapped_slice_mut⚠
Safety

mapped_sub_range
Get sub-range of memory mapping. range is in memory object space. sub is a range inside range. ptr points to the range.start offset from memory origin. returns pointer to the range.starti + sub.start offset from memory origin if sub fits in range.