Module rendy_memory::usage

Defines usage types for memory bocks. See Usage and implementations for details.

Structs

Data
Full speed GPU access. Optimal for render targets and persistent resources. Avoid memory with host access.

Download
GPU to CPU data flow with mapping. Used for copying data from Data memory to be read by the host. Host access is guaranteed.

Dynamic
CPU to GPU data flow with update commands. Used for dynamic buffer data, typically constant buffers. Host access is guaranteed. Prefers memory with fast GPU access.

Upload
CPU to GPU data flow with mapping. Used for staging data before copying to the Data memory. Host access is guaranteed.

Enums

MemoryUsageValue
Well-known memory usage types.

Traits

MemoryUsage
Memory usage trait.